Basic contract for owning Xaya accounts.
This first commit initialises the basic framework for contract
development with Truffle, and adds a first contract, AccountHolder.
This contract has the logic for owning a Xaya account (including the
initialisation by transferring one to it), and is able to send moves
with it; this will be utilised for trading vaults in subcontracts.