Implement VaultManager contract.
The VaultManager is a subcontract of AccountHolder.  It keeps track of
known vaults controlled by the owned account, and is able to send
moves for creating/funding vaults and sending funds from a vault.
The on-chain state will be kept in sync accordingly.

This is the basis for the Democrit contract, which builds the orderbook
and execution on top of the base vault logic.