Implement checkpointing.
This adds checkpointing functionality to the VaultManager.  It keeps track
of checkpoints, which e.g. Democrit can then verify from order matching.

Checkpoints can be created (saved in the contract + requested from the
GSP with a move) with a new method, maybeCreateCheckpoint.  They are also
automatically created as needed if an operation (vault creation or
sending from a vault) is done on the contract.