Main contract for Democrit.
This adds a contract for Democrit, which for now just inherits from
LimitSelling.  In the future, it will be based on LimitBuying, and will
also provide some small extras, like perhaps methods to verify the state of
buy orders for the frontend.

This contract will serve as the "entry point" for the Democrit exchange,
but does not and will not contain any main functionality itself.  Instead
it consolidates logic from parent contracts.