Explicitly escape UTF-8 in JsonUtils.
Instead of copying over bytes potentially encoding UTF-8 in the JsonUtils,
explicitly parse UTF-8 and convert the codepoints to \uXXXX escape
sequences.

While the previous approach also produces valid JSON for valid input
(and invalid JSON for invalid input), this is now more explicit, and
produces pure-ASCII output.  This is, in particular, required to pass
the XayaPolicy move check.