Define subgraph for Democrit events.
This defines a "The Graph" subgraph, which indexes events for
trades, vaults and sell orders in a Democrit contract.  The subgraph
is set up so it can be developed, built and tested, but is not
meant to be directly deployed.  Rather, projects using Democrit for
a concrete application can then use the code to deploy their own
subgraph, linked to their specific contract and network.