Tie sell orders to an address.
This records an address together with each active SellOrder, namely
the owner of the seller account at creation time.  The order is only
valid (when someone takes it) if the owner still matches the address.

With this, we prevent potential attacks and abuse when names associated to
orders are sold or transferred.