Helper method for checking sell orders.
This adds a helper method to the Democrit main contract, which can
be used to check on a batch of sell orders at once (in particular,
verifying if the creator address still matches the current account
owner, which is required for an order to be valid).

With this, frontends can quickly verify orders from the book to see
which ones to actually use or display, all in a single contract call.