Split VaultManager to be a stand-alone contract.
With all the Democrit trading features added into a single contract,
we will run into the codesize limit for contracts.  This change
makes the VaultManager a stand-alone contract, which will work along-side
the Democrit trading contracts (instead of them deriving from it).

The VaultManager becomes an Ownable, and the internal methods for creating
vaults and spending from vaults become public and onlyOwner.  The owner
is the Democrit contract, so it can continue to use the vaults as per
the protocol.  This also ensures that all remaings fully trustless and
decentralised, even with "ownership" as a concept added.