Abstract ABI loading for poolbot.
This refactors the poolbot Python library, so that the loading of
ABI files for the smart contracts it uses is abstracted (by a closure
the user of the library has to pass in).

With this, the poolbot no longer depends on / assumes that the code is
used in the context of the NPM package, but can be used more generally,
e.g. with Forge as part of a git submodule, or in a container image.