Hack to make truffle tests run.
Updates delegation-contract to the latest polygon version (which explicitly
defines forge-std to avoid potential inconsistencies), and also adds a hack
to the Truffle testutils.js file to load and support build artefacts
from forge.

Also updated Truffle to 5.11.5, which supports EIP-1193 request() used by
the test code in its provider.

With this, "./truffle test" runs, with a caveat:  The code in the two
submodules lib/polygon-contract and lib/delegation-contract needs to be built
directly there with "forge build", and the build config needs to be set
to build for "london" EVM for proper truffle support.